server/requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.16.0
Pillow>=9.0.0
Flask>=2.3.0
numpy>=1.21.0
requests>=2.28.0